requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.19
djangorestframework==3.15.2
joblib==1.4.2
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-surprise==1.1.4
scipy==1.10.1
six==1.17.0
sqlparse==0.5.3
surprise==0.1
typing-extensions==4.12.2
tzdata==2025.1
